Make last file in list visible
For some reason the CSS didn't connect the side-nav to the bottom
of the page and this lead to some wrong calculations of the browser
which lead to the fact that the last file in the list of files was
outside of the visible area.

Pinning it with bottom: 0; fixes this problem and now one can scroll
down all the way to even see the last file in the file list.