Allow for parens in entity friendly names
Selects the last encountered '('-split instead of the second, allowing for parenthesis in friendly entity names.

Fixes https://github.com/home-assistant/addons/issues/2779

A bit dirty. Actual entity names should be put as data-attributes in the elements and not be extracted from the shown text, but this patch is minimally invasive.